Separate User and PhotoProfile classes into their own files
